fix potentially implemented, however i will need to simplify it a slight bit

+++ src/client/java/com/bandw/render/ShieldBlockEntityRenderer.java
@@ -1,18 +1,21 @@
 package com.bandw.render;
 
 import com.bandw.ClientMain;
-import net.minecraft.client.render.VertexConsumer;
-import net.minecraft.client.render.model.ModelLoader;
-import net.minecraft.util.math.Direction;
-import net.minecraft.client.render.RenderLayer;
-import net.minecraft.client.render.BlockEntityRenderDispatcher;
+import net.fabricmc.api.EnvType;
+import net.fabricmc.api.Environment;
 import net.minecraft.client.render.block.entity.BlockEntityRenderer;
+import net.minecraft.client.render.block.entity.BlockEntityRendererFactory;
+import net.minecraft.client.render.model.ModelLoader;
 import net.minecraft.client.util.math.MatrixStack;
+import net.minecraft.client.render.VertexConsumer;
 import net.minecraft.client.render.VertexConsumerProvider;
+import net.minecraft.client.render.RenderLayer;
 import net.minecraft.util.Identifier;
 import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.math.MathHelper;
 import net.minecraft.util.math.Vec3d;
+import net.minecraft.util.math.MathHelper;
+import com.mojang.blaze3d.systems.RenderSystem;
+import com.mojang.math.Matrix4f;
 public class ShieldBlockEntityRenderer extends BlockEntityRenderer<ShieldBlockEntity> {
     private static final Identifier TEXTURE = Identifier.of(ClientMain.MOD_ID,"textures/entity/shield.png");
     public ShieldBlockEntityRenderer(BlockEntityRenderDispatcher dispatcher) {
@@ -20,19 +23,12 @@ public class ShieldBlockEntityRenderer extends BlockEntityRenderer<ShieldBlockEn
     };
     @Override
     public void render(ShieldBlockEntity entity, float tickDelta, MatrixStack matrices, VertexConsumerProvider vertexConsumers, int light, int overlay) {
-        // Push the current transformation matrix
         matrices.push();
-
-        // Translate to the center of the block
         BlockPos pos = entity.getPos();
         matrices.translate(pos.getX() + 0.5, pos.getY() + 0.5, pos.getZ() + 0.5);
-
-        // Scale the cube based on the shield size
-        float size = MathHelper.clamp(entity.getShield().getSize(), 1.0F, 100.0F); // Adjust the max size as needed
+        float size = MathHelper.clamp(entity.getShield().getSize(), 1.0F, 100.0F);
         matrices.scale(size, size, size);
-        // Render the cube with your texture
         renderCube(matrices, vertexConsumers, light, overlay, TEXTURE);
-        // Pop the transformation matrix
         matrices.pop();
     };
     public void renderCube(MatrixStack matrices, VertexConsumerProvider vertexConsumers, int light, int overlay, Identifier texture) {
@@ -57,7 +53,6 @@ public class ShieldBlockEntityRenderer extends BlockEntityRenderer<ShieldBlockEn
             {0, 4, 7, 3}, // Left
             {1, 5, 6, 2}  // Right
         };
-        // Corresponding UV coordinates for the vertices of each face
         float[][] uvs = {
             {0, 0, 1, 0, 1, 1, 0, 1}, // Back
             {0, 0, 1, 0, 1, 1, 0, 1}, // Front
@@ -71,16 +66,15 @@ public class ShieldBlockEntityRenderer extends BlockEntityRenderer<ShieldBlockEn
             float[] uv = uvs[faceIndex];
             for (int i = 0; i < 4; i++) {
                 Vec3d vertex = vertices[face[i]];
-                vertexConsumer.vertex(matrices.peek().getPositionMatrix(), vertex.getX(), vertex.getY(), vertex.getZ())
+                vertexConsumer.vertex(matrices.peek().getPositionMatrix(), (float) vertex.getX(), (float) vertex.getY(), (float) vertex.getZ())
                     .color(255, 255, 255, 255)
-                    .texture(uv[i * 2], uv[i * 2 + 1]) // Set the texture coordinates
+                    .texture(uv[i * 2], uv[i * 2 + 1])
                     .overlay(overlay)
                     .light(light)
                     .normal(0, 1, 0)
                     .next();
             };
         };
-        // Pop the transformation matrix
         matrices.pop();
     };
 };
